define INET6 on IRIX, making libnet compile cleanly.
Turned out to be an easy fix, once I found some time to look into it.

I'm not an expert on IRIX and I only have the one box, but the include
files (from the IRIX developer tools CDs) seemed to be referencing
some other include files that didn't exist on my system. Luckily it
turned out that what appears to be a faulty include wasn't actually
needed.

With the below patch it compiles cleanly for me, and seems like it
shouldn't break anything.