dlpi: Try harder to find the device for the interface.
Solaris usually has network devices (with and without unit numbers) in /dev,
but if a directory of the same name already exists then it will not create
any devices at all directly in /dev.

Devices are also created dynamically in /dev/net/. This patch will make
libnet look there too, and not fail if it encounteres a directory.

Example:
In my virtualised setup the network interface is named "net0", and
/dev/net/ is a directory, so /dev/net0 isn't created. /dev/net/net0 works
fine though.