Use standard autotools doxygen support.
The doxygen target directory is moved to the default location.

Generation of PDF doxygen documentation is enabled by default, can be
disabled with

$ ./configure --disable-doxygen-pdf

This uses ax_prog_doxygen.m4 which is part of the standard gnu
autoconf archive.  E.g.,

$ apt-get install autoconf-archive

A copy of /usr/share/aclocal/ax_prog_doxygen.m4 is put in m4/.

This is invoked in configure.ac, and requires a macro in Makefile.am