libnet.doxygen.conf is now libnet.doxygen.conf.in
It's not "really" needed. autoconf does nothing with it (yet). It's just to give a hint that it'll only generate docs during process never with standalone doxygen. That's because a lot of fields had to be replaced with variables that are set and determined during configure time. Either by user (generate docs or not) or by the various tests performed by ax_prog_doxygen.m4.
it's also updated to doxygen 1.8.14 and all comments have been stripped out (doxygen -s -u).