Fix number ranges produced by libnet_get_prand()
Modulo X gives a range of 0..(X-1).  LIBNET_PR8, LIBNET_PR16,
LIBNET_PRu16 and LIBNET_PR32 all were off by one and never produced the
highest number (e.g. 0xff for LIBNET_PR8).

Since all of the number ranges are powers of 2, we can mask instead of
modulo, which should be slightly faster; now it makes sense to use the
highest number we want to produce as a mask.  Use masking also for
LIBNET_PR2; this produces the same result but is consistent with the
other ranges.