Update ChangeLog for v1.2 release
This is a *major* cleanup.  All fixes and changes made to build system;
warnings/errors, cosmetic doc fixes, and other changes not relevant to
end users have been removed and/or rephrased.

I've also removed the names os people doing the changes, as this also
does not really provide any value to an end user but is more of a mix
with a CONTRIB/AUTHORS file, we have GIT if attribution/blame is needed
check there.   Otherwise the curator(s) are to blame.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>