Remove pre-built binary files from win32/wpdpack
Not only is it our policy to not include binaries in GIT, the wpdpack
distribution comes under another license, see #85, which means libnet
top LICENSE file would otherwise have to be changed.

We recommend installing https://nmap.org/npcap/ for using libnet on
Windows.  For cross-building on a non-windows host, see README.win32.

Closes #93

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>