Update local 4-clause BSD getopt.c with FreeBSD 3-clause version
This patch updates the local getopt.c, used only when building samples
on Windows, with a newer version with better license from FreeBSD.

For reference, Berkeley recinded the 3rd clause of the original 4-clause
license in 1999.  This was adopted by the BSD UNIX community, and later
the numbering was also updated.

https://bsd.slashdot.org/story/99/09/02/189210/berkeley-removes-advertising-clause

For some reason, the replacement header file, win32/getopt.h, comes with
the 2-clause BSD license.  I did no more research on that, but I guess
the FreeBSD VCS has the complete story.

Related to issue #85.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>