Make man-page generation reproducible
For this we behave similar to pod2man and use the input file mtime.

Also adapt date format to what doxygen-1.8.20 produces
to make the comment above true again.

LC_ALL=C is used so that %B is independent of language.

Also use UTC to be independent of timezone.

This patch was done while working on reproducible builds for openSUSE.