Drop Travis-CI in favor of GitHub Actions
The new gh action builder now fully replaces Travis-CI, so we can
finally drop it.  They stopped supporting builds on travis.org and
did not provide an automatic migration option, so this was a good
opportunity to centralize on GitHub.  GH Actions are really quick
and allows for a lot of interesting possiblities, e.g. automated
testing using dummy nic in a docker container (see Nemesis).

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>