libnet_if_addr.c: fix segfault when number of IPs > 512
The current code has a hardcoded limit of 512 IP addresses.
The static array will be overflowed when more addresses are
present.

Fix the static limit by using realloc when the array needs to grow
to accomodate more interfaces.

Growing the array by a factor of 1.5 each time (close to golden ratio)
to maximize reuse of previously allocated space.