libnet_if_addr.c: fix 'Using uninitialized value "rc".'
This fixes static code analysis report:

  1. libnet-1.3/src/libnet_if_addr.c:551:5: var_decl: Declaring variable "rc" without initializer.
  8. libnet-1.3/src/libnet_if_addr.c:626:5: uninit_use: Using uninitialized value "rc".
  #   624|       }
  #   625|
  #   626|->     return rc;
  #   627|   }
  #   628|

The code was jumping to the 'end' label without setting rc to anything.
Doing 'return rc' will indeed return an uninitialized value for some
cases.

This commit removed the 'bad' label and in an error case always jumps to
'end' with rc initialized to -1.

Signed-off-by: Adrian Reber <areber@redhat.com>